pyproject.toml
[project]
name = "mcp-server-sqlserver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aioodbc>=0.5.0",
    "fastmcp>=2.3.3",
    "greenlet>=3.2.1",
    "httpx>=0.28.1",
    "mcp[cli]>=1.8.0",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.40",
    "starlette>=0.46.2",
]
